EECS_430_Team4-main/python/requirements.txt
pyserial
opencv-python
numpy
matplotlib